Module sample::conv

Pure functions and traits for converting between i8, i16, I24, i32, I48, i64, u8, u16, U24, u32, U48, u64, f32 and f64.

Each conversion function is performance focused, memory-sensitive and expects that the user has validated their input prior to the function call.

No conversion function will ever cast to a type with a size in bytes larger than the largest between the source and target sample types.

The conversion functions do not check the range of incoming values for floating point values or any of the custom I24, U24, I48 and U48 types.

Modules

f32

f64

i16

i24

i32

i48

i64

i8

u16

u24

u32

u48

u64

u8

Traits

Duplex
Sample types which may be converted to and from some type S.

DuplexBoxedFrameSlice

DuplexBoxedSampleSlice

DuplexBoxedSlice

DuplexFrameSlice

DuplexFrameSliceMut

DuplexSampleSlice

DuplexSampleSliceMut

DuplexSlice

DuplexSliceMut

FromBoxedFrameSlice
For converting from a boxed slice of Frames to a boxed slice of Samples.

FromBoxedSampleSlice
For converting a boxed slice of Samples to a boxed slice of Frames.

FromFrameSlice
For converting from a slice of Frames to a slice of Samples.

FromFrameSliceMut
For converting from a slice of Frames to a slice of Samples.

FromSample
Similar to the std From trait, but specifically for converting between sample types.

FromSampleSlice
For converting from a slice of Samples to a slice of Frames.

FromSampleSliceMut
For converting from a mutable slice of Samples to a mutable slice of Frames.

ToBoxedFrameSlice
For converting from a boxed slice of Samples to a boxed slice of Frames.

ToBoxedSampleSlice
For converting from a boxed slice of Frames to a boxed slice of Samples.

ToFrameSlice
For converting from a slice of Samples to a slice of Frames.

ToFrameSliceMut
For converting from a mutable slice of Samples to a mutable slice of Frames.

ToSample
Similar to the std Into trait, but specifically for converting between sample types.

ToSampleSlice
For converting from a slice of Frames to a slice of Samples.

ToSampleSliceMut
For converting from a mutable slice of Frames to a mutable slice of Samples.