Module sample::frame

Use the Frame trait to remain generic over the number of channels at a single discrete moment in time.

Implementations are provided for all fixed-size arrays up to 32 elements in length.

Structs

Channels
An iterator that yields the sample for each channel in the frame by value.

N1
A typified version of a number of channels.

N10
A typified version of a number of channels.

N11
A typified version of a number of channels.

N12
A typified version of a number of channels.

N13
A typified version of a number of channels.

N14
A typified version of a number of channels.

N15
A typified version of a number of channels.

N16
A typified version of a number of channels.

N17
A typified version of a number of channels.

N18
A typified version of a number of channels.

N19
A typified version of a number of channels.

N2
A typified version of a number of channels.

N20
A typified version of a number of channels.

N21
A typified version of a number of channels.

N22
A typified version of a number of channels.

N23
A typified version of a number of channels.

N24
A typified version of a number of channels.

N25
A typified version of a number of channels.

N26
A typified version of a number of channels.

N27
A typified version of a number of channels.

N28
A typified version of a number of channels.

N29
A typified version of a number of channels.

N3
A typified version of a number of channels.

N30
A typified version of a number of channels.

N31
A typified version of a number of channels.

N32
A typified version of a number of channels.

N4
A typified version of a number of channels.

N5
A typified version of a number of channels.

N6
A typified version of a number of channels.

N7
A typified version of a number of channels.

N8
A typified version of a number of channels.

N9
A typified version of a number of channels.

Traits

Frame
Represents one sample from each channel at a single discrete instance in time within a PCM signal.

NumChannels
Restricts the types that may be used as the Frame::NumChannels associated type.

Type Definitions

Mono

Stereo