Trait sample::frame::NumChannels

pub trait NumChannels { }

Restricts the types that may be used as the Frame::NumChannels associated type.

NumChannels allows us to enforce the number of channels that a Frame must have in certain operations. This is particularly useful for Frame::map and Frame::zip_map, as it allows us to guarantee that the input and output frame types will retain the same number of channels at compile-time, and in turn removes the need for bounds checking.

This trait is implemented for types N1...N32.