Module sample::interpolate

The Interpolate module allows for conversion between various sample rates.

Structs

Converter
An iterator that converts the rate at which frames are yielded from some given frame Interpolator into a new type.

Floor
Interpolator that just rounds off any values to the previous value from the source

Linear
Interpolator that interpolates linearly between the previous value and the next value

Sinc
Interpolator for sinc interpolation. Generally accepted as one of the better sample rate converters, although it uses significantly more computation.

Traits

Interpolator
Types that can interpolate between two values.