Struct sample::interpolate::Floor

pub struct Floor<F> { /* fields omitted */ }

Interpolator that just rounds off any values to the previous value from the source

Methods

impl<F> Floor<F>

fn new(left: F) -> Floor<F>

Create a new Floor Interpolator.

fn from_source<S>(source: &mut S) -> Floor<F> where
F: Frame,
S: Signal<Frame = F>,

Consumes the first value from a given source in order to initialize itself. If the source has no values at all, this will return None.