Struct sample::interpolate::Sinc

pub struct Sinc<F> { /* fields omitted */ }

Interpolator for sinc interpolation. Generally accepted as one of the better sample rate converters, although it uses significantly more computation.

Methods

impl<F> Sinc<F>

fn new<S>(depth: usize, padding: S) -> Self where
F: Frame,
S: Signal<Frame = F>,

Create a new Sinc interpolater whose inner queue will be padded with the given signal.

fn zero_padded(depth: usize) -> Self where
F: Frame,

Create a new Sinc interpolator whose inner queue will be padded with equilibrium frames.