Struct sample::signal::Bus

pub struct Bus<S> where S: Signal, S::Item: Frame {
    // some fields omitted
}

A type which allows for sending a single Signal to multiple outputs.

This type manages

Methods

impl<S> Bus<S> where S: Signal, S::Item: Frame

fn send(&self) -> Output<S>

Produce a new Output node to which the signal S will output its frames.