Struct sample::signal::IntoInterleavedSamples

pub struct IntoInterleavedSamples<S> where
    S: Signal,  { /* fields omitted */ }

Converts a Signal to a type that yields the individual interleaved samples.

Methods

impl<S> IntoInterleavedSamples<S> where
S: Signal,

fn next_sample(&mut self) -> <S::Frame as Frame>::Sample

Yield the next interleaved sample from the inner Signal.

fn into_iter(self) -> IntoInterleavedSamplesIterator<S>

Convert the ToInterleavedSamples into an Iterator.