Struct sample::signal::Rate

pub struct Rate {
    // some fields omitted
}

The rate at which phrase a Signal is sampled.

Methods

impl Rate

fn const_hz(self, hz: f64) -> ConstHz

Create a ConstHz iterator which consistently yields "hz / rate".

fn hz<I>(self, init: f64, hz: I) -> Hz<I> where I: Iterator<Item=f64>

Create a variable hz some iterator that yields hz and an initial hz.

The Hz iterator yields phase step sizes equal to "hz / rate".