Trait sample::FloatSample

pub trait FloatSample: Sample<Signed=Self, Float=Self> + SignedSample + Mul<Output=Self> + Div<Output=Self> + Duplex<f32> + Duplex<f64> {
    fn identity() -> Self;
    fn sample_sqrt(self) -> Self;
}

Sample format types represented as floating point numbers.

Samples often need to be converted to some mutual FloatSample type for signal scaling and modulation.

Required Methods

fn identity() -> Self

Represents the multiplicative identity of the floating point signal.

fn sample_sqrt(self) -> Self

Calculate the square root of Self. A convenience generic wrapper around .sqrt().