Struct sample::types::i24::I24

pub struct I24(_);

Methods

impl I24

fn new(val: i32) -> Option<Self>

Construct a new sample if the given value is within range.

Returns None if val is out of range.

fn new_unchecked(s: i32) -> Self

Constructs a new sample without checking for overflowing.

This should only be used if the user can guarantee the sample will be within range and they require the extra performance.

If this function is used, the sample crate can't guarantee that the returned sample or any interacting samples will remain within their MIN and MAX bounds.

fn inner(self) -> i32

Return the internal value used to represent the sample type.