Module sample::window

Module for windowing over a batch of Frames. Includes default Hanning and Rectangle window types.

Structs

Hanning
A type of window function, also known as teh "raised cosine window".

Rectangle
The simplest window type, equivalent to replacing all but N values of data sequence by zeroes, making it appear as though the waveform suddenly turns on and off.

Window
A Signal type that for every yielded phase, yields the amplitude across the window::Type for that phase.

Windowed
An Iterator that multiplies a Signal with a Window.

Windower
Takes a long slice of frames and yields Windowed chunks of size bin once every hop frames.

Traits

Type
The window function used within a Window.

Functions

hanning
A helper function for constructing a Window that uses a Hanning Type function.

rectangle
A helper function for constructing a Window that uses a Rectangle Type function.